Support non-consumer connections, fixes #36
A consumer is a Worker process that sends a non-empty WID in its HELLO.  Pure producers do not have to send a WID with HELLO and do not need to BEAT.

Despite this distinction, Consumers can still FETCH and ACK.  I decided to be liberal here because of things like test suite runners that might not call themselves “workers” but still want to connect to Faktory and run commands.

Only consumers should BEAT and are displayed on the Busy page.