Remove TLS server support
There are many other ways to add encryption which are orthogonal to Faktory.
It is not supported to have Faktory listening on the open Internet.

- Use stunnel/spiped with Faktory inside a container
- Set up an SSH tunnel between machines

It’s possible we will add back in direct TLS support but not until we get a lot more production feedback from users so we have a clear real world direction and not a vague sense.