Enable running Docker Faktory server with password
This allows using `make drun` to start the server with a password
without editing the Makefile. Specifically, you can now do:

```
make PASSWORD=test drun
```

And the server will run with the password `test`. The old invocation of

```
make drun
```

Will run without a password. This is because the server binary
determines whether a password is set or not by whether or not it is the
empty string:
https://github.com/contribsys/faktory/blob/00eb4f6c/server/security.go#L30