Add separate target for loadtest
loadtest is a client-only binary, and thus does not need any of the
Faktory dependencies (e.g., webui or rocksdb). This makes it much easier
for developers to load test a Faktory server run with `make drun` by not
forcing them to also sort out and build all the dependencies.