Repackage core API into client
We want Go clients to be able to pull only the code they need to build a Faktory client.  Currently the core API is in the root “faktory” package, which means Go wants to build everything beneath it.  Since clients don’t need all of that code, they shouldn’t have to download and build it just to access the Client and Job code.