Overhaul of counter types
Most counters in the system are int64, as that is the default Go integer type.  However counters are not meant to represent negative numbers and we want to actively fight against negative values.  Queue#size was overhauled to use uint months ago but SortedSet#size didn’t get the same refactor until now.

This refactor is part 1 of fixing #148.