Update Faktory error handling, fixes #208
Introduce new concept of "KnownError", an error with a code which can be programmatically acted on by the Faktory client.  An example is when the client pushes a non-unique job, how should the Unique Job feature respond?

Previously it returned this:

-ERR Job is not unique

But "ERR" is a nebulous code and we don't want people building logic on the message.

Now the feature will return an error in RESP that looks like this:

-NOTUNIQUE Job is not unique

Generic and unexpected errors still will return ERR for the code.  It is recommended that client libraries raise an exception on ERR, otherwise return the code to the caller.